src/fastapi_with_mcp/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
langchain==0.3.21
langchain-mcp-adapters==0.0.5
langchain-openai==0.3.11
langgraph-api==0.0.38
langgraph-cli==0.1.81
pip==25.0.1
xmltodict==0.14.2
